florence-2/requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
spacy>=3.7.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0 